fix: OverdueRate double->decimal for Dapper; fix angular.json serve buildTarget ref

+++ backend/src/Modules/Reporting/STRIDE.Modules.Reporting.Application/ReadModels/CompletionTimeAnalyticsDto.cs
@@ -44,5 +44,5 @@ public sealed record TeamMemberPerformanceDto(
     string Role,
     int    CompletedSteps,
     int    CompletedWorkflows,
-    double AvgStepDurationMinutes,
-    double OverdueRate);
+    double  AvgStepDurationMinutes,
+    decimal OverdueRate);
